Change floats in PTYTextView to doubles. Fixes bug 1029.

+++ Headers/iTerm/PTYTextView.h
@@ -48,7 +48,7 @@ struct PTYFontInfo {
     NSFont* font;  // Toll-free bridged to CTFontRef
 
     // Metrics
-    float baselineOffset;
+    double baselineOffset;
 
     struct PTYFontInfo* boldVersion;  // may be NULL
 };
@@ -90,12 +90,12 @@ typedef struct PTYFontInfo PTYFontInfo;
     BOOL colorInvertedCursor;
 
     // geometry
-    float lineHeight;
-    float lineWidth;
-    float charWidth;
-    float charWidthWithoutSpacing, charHeightWithoutSpacing;
-    float horizontalSpacing_;
-    float  verticalSpacing_;
+    double lineHeight;
+    double lineWidth;
+    double charWidth;
+    double charWidthWithoutSpacing, charHeightWithoutSpacing;
+    double horizontalSpacing_;
+    double  verticalSpacing_;
 
     PTYFontInfo primaryFont;
     PTYFontInfo secondaryFont;
@@ -110,7 +110,7 @@ typedef struct PTYFontInfo PTYFontInfo;
     NSColor* cursorTextColor;
 
     // transparency
-    float transparency;
+    double transparency;
 
     // data source
     VT100Screen *dataSource;
@@ -164,7 +164,7 @@ typedef struct PTYFontInfo PTYFontInfo;
 
     // Scrolls view when you drag a selection to top or bottom of view.
     NSTimer* selectionScrollTimer;
-    float prevScrollDelay;
+    double prevScrollDelay;
     int scrollingX;
     int scrollingY;
     NSPoint scrollingLocation;
@@ -179,8 +179,8 @@ typedef struct PTYFontInfo PTYFontInfo;
     int accY;
 
     BOOL advancedFontRendering;
-    float strokeThickness;
-    float minimumContrast_;
+    double strokeThickness;
+    double minimumContrast_;
 
     BOOL changedSinceLastExpose_;
 
@@ -253,7 +253,7 @@ typedef struct PTYFontInfo PTYFontInfo;
 }
 
 + (NSCursor *)textViewCursor;
-+ (NSSize)charSizeForFont:(NSFont*)aFont horizontalSpacing:(float)hspace verticalSpacing:(float)vspace;
++ (NSSize)charSizeForFont:(NSFont*)aFont horizontalSpacing:(double)hspace verticalSpacing:(double)vspace;
 - (id)initWithFrame:(NSRect)aRect;
 - (void)dealloc;
 - (BOOL)becomeFirstResponder;
@@ -302,7 +302,7 @@ typedef struct PTYFontInfo PTYFontInfo;
 //get/set methods
 - (NSFont *)font;
 - (NSFont *)nafont;
-- (void)setFont:(NSFont*)aFont nafont:(NSFont *)naFont horizontalSpacing:(float)horizontalSpacing verticalSpacing:(float)verticalSpacing;
+- (void)setFont:(NSFont*)aFont nafont:(NSFont *)naFont horizontalSpacing:(double)horizontalSpacing verticalSpacing:(double)verticalSpacing;
 - (NSRect)scrollViewContentSize;
 - (void)setAntiAlias:(BOOL)asciiAA nonAscii:(BOOL)nonAsciiAA;
 - (BOOL)useBoldFont;
@@ -337,7 +337,7 @@ typedef struct PTYFontInfo PTYFontInfo;
 - (int)selectionEndY;
 - (void)setSelectionFromX:(int)fromX fromY:(int)fromY toX:(int)toX toY:(int)toY;
 
-- (float)excess;
+- (double)excess;
 
 
 - (NSDictionary*)markedTextAttributes;
@@ -347,10 +347,10 @@ typedef struct PTYFontInfo PTYFontInfo;
 - (void)setDataSource:(id)aDataSource;
 - (id)delegate;
 - (void)setDelegate:(id)delegate;
-- (float)lineHeight;
-- (void)setLineHeight:(float)aLineHeight;
-- (float)charWidth;
-- (void)setCharWidth:(float)width;
+- (double)lineHeight;
+- (void)setLineHeight:(double)aLineHeight;
+- (double)charWidth;
+- (void)setCharWidth:(double)width;
 
 // Update the scroll position and schedule a redraw. Returns true if anything
 // onscreen is blinking.
@@ -364,15 +364,15 @@ typedef struct PTYFontInfo PTYFontInfo;
 - (void)deselect;
 
 // transparency
-- (float)transparency;
-- (void)setTransparency:(float)fVal;
+- (double)transparency;
+- (void)setTransparency:(double)fVal;
 - (BOOL)useTransparency;
 
 - (void)setSmartCursorColor:(BOOL)value;
-- (void)setMinimumContrast:(float)value;
+- (void)setMinimumContrast:(double)value;
 
 // Dim all colors towards gray
-- (void)setDimmingAmount:(float)value;
+- (void)setDimmingAmount:(double)value;
 
 //
 // Drag and Drop methods for our text view
@@ -434,8 +434,8 @@ typedef struct PTYFontInfo PTYFontInfo;
 - (BOOL)readSelectionFromPasteboard:(NSPasteboard *)pboard;
 - (BOOL)findInProgress;
 - (BOOL)continueFind;
-- (float)horizontalSpacing;
-- (float)verticalSpacing;
+- (double)horizontalSpacing;
+- (double)verticalSpacing;
 
 // This textview is about to become invisible because another tab is selected.
 - (void)aboutToHide;
@@ -494,15 +494,15 @@ typedef enum {
 
 - (NSString *)_getURLForX:(int)x y:(int)y;
 // Returns true if any char in the line is blinking.
-- (BOOL)_drawLine:(int)line AtY:(float)curY toPoint:(NSPoint*)toPoint;
+- (BOOL)_drawLine:(int)line AtY:(double)curY toPoint:(NSPoint*)toPoint;
 - (void)_drawCursor;
 - (void)_drawCursorTo:(NSPoint*)toOrigin;
 - (void)_drawCharacter:(screen_char_t)screenChar
                fgColor:(int)fgColor
     alternateSemantics:(BOOL)fgAlt
                 fgBold:(BOOL)fgBold
-                   AtX:(float)X
-                     Y:(float)Y
+                   AtX:(double)X
+                     Y:(double)Y
            doubleWidth:(BOOL)double_width
          overrideColor:(NSColor*)overrideColor;
 
@@ -543,14 +543,14 @@ typedef enum {
 - (void)invalidateInputMethodEditorRect;
 
 // Return the number of pixels tall to draw the cursor.
-- (float)cursorHeight;
+- (double)cursorHeight;
 
 // Draw the contents of the input method editor beginning at some location, 
 // usually the cursor position.
 // xStart, yStart: cell coordinates
 // width, height: cell width, height of screen
 // cursorHeight: cursor height in pixels
-- (BOOL)drawInputMethodEditorTextAt:(int)xStart y:(int)yStart width:(int)width height:(int)height cursorHeight:(float)cursorHeight;
+- (BOOL)drawInputMethodEditorTextAt:(int)xStart y:(int)yStart width:(int)width height:(int)height cursorHeight:(double)cursorHeight;
 
 - (BOOL)_wasAnyCharSelected;
